Convert shipping/payment method id to integer
This is necessary to correctly handle parameters that are being sent in a wrong way by certain older lsjs versions (e.g. 2.0.11). Because of a double question mark in the url, the method id parameter was interpreted as e.g. "id=2?cajaxCall=1668086864059". In PHP 7 Merconis somehow managed to eventually set the method id to the correct value (e.g. 2) but in PHP the method id was actually processed as "2?cajaxCall=1668086864059". The error is originally triggered by a bug in lsjs.js:L61 in lsjs 2.0.11 and older.