Improve insert tag "shopPicture"
If the size parameter is present in the shopPicture insert tag but there's no image size record with the specified alias, the size parameter gets removed from the parameter string. After removing the size parameter the parameter string could sometimes contain leftover "?" or "&" characters that could later lead to a warning when Contao rendered the picture insert tag with the incorrect parameter string.